Update invoice_data_extraction.py
I wrote this code for extracting details from amazon.in invoices. 
This code is written assuming that the inputs are at "A:\Projects\Invoice Data Extraction\Inputs" location and all the outputs will be created in "A:\Projects\Invoice Data Extraction\Outputs" i.e,  one xlsx and one Json file for each and every invoice and also one xlsx and one Json file where they contain details of all the invoices.
I extracted the following details 
1. Order date
2. Order Number
3. Billed amount
4. Gst number
5. Biller name

I hope that my code works for any invoice of amazon.in for most of the cases.